requirements.txt
pathlib==1.0.1
pillow==11.2.1
setuptools==78.1.1
wheel==0.45.1
